Extract TS reranking to its own plugin
