+++ x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/XPackFeatures.java
@@ -12,7 +12,6 @@ import org.elasticsearch.features.FeatureSpecification;
 import org.elasticsearch.features.NodeFeature;
 import org.elasticsearch.license.License;
 import org.elasticsearch.xpack.core.datatiers.NodesDataTiersUsageTransportAction;
-import org.elasticsearch.xpack.core.ml.rerank.TextSimilarityRankRetrieverBuilder;
 
 import java.util.Map;
 import java.util.Set;
@@ -26,8 +25,7 @@ public class XPackFeatures implements FeatureSpecification {
     public Set<NodeFeature> getFeatures() {
         return Set.of(
             NodesDataTiersUsageTransportAction.LOCALLY_PRECALCULATED_STATS_FEATURE, // Added in 8.12
-            License.INDEPENDENT_TRIAL_VERSION_FEATURE, // 8.14.0
-            TextSimilarityRankRetrieverBuilder.NODE_FEATURE
+            License.INDEPENDENT_TRIAL_VERSION_FEATURE // 8.14.0
         );
     }
