+++ x-pack/plugin/rank-rrf/src/main/java/org/elasticsearch/xpack/rank/rrf/RRFRetrieverBuilder.java
@@ -60,6 +60,11 @@ public final class RRFRetrieverBuilder extends RetrieverBuilder {
         RetrieverBuilder.declareBaseParserFields(NAME, PARSER);
     }
 
+    // - new plugin
+    // - feature support check
+    // - licence check
+    // - REST tests
+    // - explain
     public static RRFRetrieverBuilder fromXContent(XContentParser parser, RetrieverParserContext context) throws IOException {
         if (context.clusterSupportsFeature(RRF_RETRIEVER_SUPPORTED) == false) {
             throw new ParsingException(parser.getTokenLocation(), "unknown retriever [" + NAME + "]");
