+++ x-pack/plugin/rank-text-similarity/build.gradle
@@ -1,0 +1,31 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+apply plugin: 'elasticsearch.internal-es-plugin'
+apply plugin: 'elasticsearch.internal-cluster-test'
+apply plugin: 'elasticsearch.internal-yaml-rest-test'
+
+esplugin {
+  name 'rank-text-similarity'
+  description 'Reciprocal rank fusion in search.'
+  classname 'org.elasticsearch.xpack.rank.textsimilarity.TextSimilarityRankPlugin'
+  extendedPlugins = ['x-pack-core', 'x-pack-inference']
+}
+
+dependencies {
+//  compileOnly project(path: xpackModule('core'))
+//  compileOnly project(path: xpackModule('inference'))
+  compileOnly project(':x-pack:plugin:core')
+  compileOnly project(':x-pack:plugin:inference')
+  compileOnly project(':x-pack:plugin:ml')
+
+  testImplementation(testArtifact(project(xpackModule('core'))))
+  testImplementation project(path: xpackModule('inference'))
+
+//  clusterModules project(xpackModule('rank-rrf'))
+//  clusterModules project(':modules:lang-painless')
+}
