+++ x-pack/plugin/rank-text-similarity/src/main/java/module-info.java
@@ -1,0 +1,20 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import org.elasticsearch.xpack.rank.textsimilarity.TextSimilarityRankFeatures;
+
+module org.elasticsearch.rank.textsimilarity {
+    requires org.apache.lucene.core;
+    requires org.elasticsearch.base;
+    requires org.elasticsearch.xcontent;
+    requires org.elasticsearch.server;
+    requires org.elasticsearch.xcore;
+
+    exports org.elasticsearch.xpack.rank.textsimilarity;
+
+    provides org.elasticsearch.features.FeatureSpecification with TextSimilarityRankFeatures;
+}
