+++ x-pack/plugin/rank-text-similarity/src/main/java/org/elasticsearch/xpack/rank/textsimilarity/TextSimilarityRankRetrieverBuilder.java
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-package org.elasticsearch.xpack.core.ml.rerank;
+package org.elasticsearch.xpack.rank.textsimilarity;
 
 import org.elasticsearch.common.ParsingException;
 import org.elasticsearch.features.NodeFeature;
